backend/pyproject.toml
[project]
name = "astralagent"
version = "0.1.0"
description = "Astral Agent - 企业级多智能体平台"
requires-python = ">=3.12"
dependencies = [
    # Web 框架
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    
    # 配置管理
    "python-dotenv>=1.0.0",
    "toml>=0.10.2",
    "pyyaml>=6.0.0",
    
    # 数据库
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.12.0",
    
    # 缓存
    "redis>=5.0.0",
    
    # 日志
    "structlog>=23.2.0",
    "python-json-logger>=2.0.0",
    
    # 监控
    "psutil>=5.9.0",
    
    # HTTP 客户端
    "httpx>=0.25.0",
    
    # 工具
    "tenacity>=8.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
]
